Refactor stack into own module
